+++ advanced-responsive-video-embedder/php/Admin/fn-settings-page.php
@@ -1,4 +1,7 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types = 1);
+
 namespace Nextgenthemes\ARVE\Admin;
 
 use Nextgenthemes\ARVE;
@@ -24,7 +27,7 @@ function settings_content(): void {
 						'title'  => true,
 					),
 				),
-				array( 'http', 'https' )
+				array( 'https' )
 			);
 			?>
 		</p>
@@ -38,14 +41,20 @@ function settings_content(): void {
 
 	<div data-wp-bind--hidden="!context.activeTabs.debug">
 
+		<?php
+		printf(
+			'<p><a href="%s">%s</a></p>',
+			esc_url( admin_url( 'site-health.php?tab=debug' ) ),
+			esc_html__( 'Site Health Debug Info', 'advanced-responsive-video-embedder' )
+		);
+		?>
+
 		<p>
-			<button data-wp-on--click="actions.deleteOembedCache" class="button-primary" style="margin-inline-end: 1em;">
+			<button data-wp-on--click="actions.deleteOembedCache" class="button-secondary" style="margin-inline-end: 1em;">
 				<?php esc_html_e( 'Delete oEmbed Cache', 'advanced-responsive-video-embedder' ); ?>
 			</button>
 			<span x-text="message"></span>
 		</p>
-
-		<?php require_once __DIR__ . '/partials/debug-info-textarea.php'; ?>
 	</div>
 
 	<?php if ( ! is_plugin_active( 'arve-pro/arve-pro.php' ) ) : ?>
@@ -86,7 +95,7 @@ function pro_message( string $addon_name, string $slug ): string {
 			'strong' => array(),
 			'a'      => array( 'href' => true ),
 		),
-		array( 'http', 'https' )
+		array( 'https' )
 	);
 }
 
@@ -97,7 +106,7 @@ function get_addon_link( string $addon_name, string $slug ): string {
 
 function settings_sidebar(): void {
 
-	if ( ! current_user_can('install_plugins') ) {
+	if ( ! current_user_can( 'install_plugins' ) ) {
 		echo '<div class="ngt-sidebar-box">';
 		esc_html_e( 'Note that you are logged in with a user who that can\'t install plugins, ask someone who can if you are interested in ARVE Extensions.', 'advanced-responsive-video-embedder' );
 		echo '</div>';
@@ -123,7 +132,7 @@ function settings_sidebar(): void {
 function print_settings_box_html( string $file ): void {
 	echo '<div class="ngt-sidebar-box">';
 	// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents
-	echo wp_kses( file_get_contents( __DIR__ . $file, false ), ARVE\ALLOWED_HTML );
+	echo wp_kses( file_get_contents( __DIR__ . $file, false ), ARVE\ALLOWED_HTML, array( 'https' ) );
 	echo '</div>';
 }