+++ advanced-responsive-video-embedder/php/Base.php
@@ -1,7 +1,9 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types = 1);
+
 namespace Nextgenthemes\ARVE;
 
-use WP_Error;
 use Nextgenthemes\WP\Settings;
 
 class Base {
@@ -9,21 +11,17 @@ class Base {
 	private Settings $settings_instance;
 	private \WP_Error $errors;
 
-	/** @readonly */
-	private array $settings_data;
-
 	public function __construct() {
-		$this->settings_data     = settings_data();
 		$this->errors            = new \WP_Error();
 		$this->settings_instance = new Settings(
 			array(
 				'namespace'           => __NAMESPACE__,
-				'settings'            => settings( 'settings_page', $this->settings_data ),
+				'settings'            => settings( 'settings_page' ),
 				'sections'            => settings_sections(),
-				'premium_sections'    => PREMIUM_SECTIONS,
-				'premium_url_prefix'  => PREMIUM_URL_PREFIX,
+				'tabs'                => settings_tabs(),
 				'menu_title'          => __( 'ARVE', 'advanced-responsive-video-embedder' ),
 				'settings_page_title' => __( 'ARVE Settings', 'advanced-responsive-video-embedder' ),
+				'plugin_file'         => PLUGIN_FILE,
 				'base_url'            => plugins_url( '', PLUGIN_FILE ),
 				'base_path'           => PLUGIN_DIR,
 			)
@@ -32,10 +30,6 @@ class Base {
 
 	public function get_settings_instance(): Settings {
 		return $this->settings_instance;
-	}
-
-	public function get_settings_data(): array {
-		return $this->settings_data;
 	}
 
 	public function get_errors(): \WP_Error {