+++ advanced-responsive-video-embedder/php/fn-oembed.php
@@ -1,4 +1,7 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types = 1);
+
 namespace Nextgenthemes\ARVE;
 
 use function Nextgenthemes\WP\valid_url;
@@ -28,7 +31,7 @@ function add_oembed_providers(): void {
 function filter_oembed_dataparse( string $html, object $data, string $url ): string {
 
 	// this is to fix Divi endless reload issue.
-	if ( is_admin() && function_exists('et_setup_theme') ) {
+	if ( is_admin() && function_exists( 'et_setup_theme' ) ) {
 		return $html;
 	}