+++ advanced-responsive-video-embedder/changelog.md
@@ -3,6 +3,12 @@
 * [ARVE Pro changelog](https://nextgenthemes.com/plugins/arve-pro/#changelog)
 * [ARVE Random Videos changelog](https://nextgenthemes.com/plugins/arve-random-video/#changelog)
 
+### 2024-01-02 10.5.0 ###
+
+* New: Debug option to control src mismatch errors.
+* Improved: Removed the debug into in favor or adding data to the Site Health Info screen.
+* Improved: Introduced a `SettingsData` and `SettingValidator` classes to make the code more robust self testing.
+
 ### 2024-12-05 10.4.0 ###
 
 * New: Added `credentialless="false"` parameter that can be used to remove the same named attribute from the iframe.